refactor(group-quiz): extract session realtime core into useGroupQuizSession
Move session/participants/loading/error state, the load-by-code effect,
the realtime channel (session patches + participant join/leave toasts),
the resolved me participant, and the host/submission derivations +
leave-guard verbatim into _hooks/use-group-quiz-session.ts (owns its own
auth + GroupQuiz.session i18n). Pruned now-unused useLeaveGuard/
guestParticipantId/supabase/useMemo/QuizParticipant/QuizSession imports.
Behavior-identical; page.tsx 1288->1182 LOC.
